refactor: split cli.ts (1141 LOC) into focused command modules, closes #44
- Extract src/cli/output.ts: logSuccess/logError/logWarning/logInfo/CLIError/handleCommandError (~58 LOC)
- Extract src/cli/commands/run.ts: run command action handler (~153 LOC)
- Extract src/cli/commands/watch.ts: watch command action handler (~156 LOC)
- Extract src/cli/commands/validate.ts: validate command action handler (~133 LOC)
- Extract src/cli/commands/list.ts: list command action handler (~107 LOC)
- Extract src/cli/commands/init.ts: init command action handler (~121 LOC)
- Extract src/cli/commands/init-templates.ts: template data for init command (~392 LOC of string data)
- Extract src/cli/commands/help.ts: help command action handler (~42 LOC)
- Slim src/cli.ts to Commander registration only (~103 LOC)

Each command module uses registerXxxCommand(program) pattern.
All commands verified working: --help, run --help, validate --help, list --help, help.
TypeScript compiles cleanly; pre-existing ProcessLifecycleManager test failure unaffected.
