+++ src/cli.ts
@@ -2,20 +2,21 @@
 
 /**
  * Command Line Interface for the Agentic Testing System
- * Enhanced with config loading, file watching, project initialization,
- * better error handling, progress indicators, and environment support.
+ *
+ * This file is the Commander entry point only. All command logic lives in
+ * src/cli/commands/. Shared output utilities are in src/cli/output.ts.
  */
 
 import { Command } from 'commander';
-import { ScenarioLoader } from './scenarios';
-import { logger, ConfigManager, createYamlParser } from './utils';
-import * as path from 'path';
-import * as fs from 'fs/promises';
-import * as chokidar from 'chokidar';
-import chalk from 'chalk';
 import * as dotenv from 'dotenv';
-import { SingleBar, Presets } from 'cli-progress';
-import { createDefaultConfig } from './lib';
+import chalk from 'chalk';
+import { logError, logInfo, logWarning } from './cli/output';
+import { registerRunCommand } from './cli/commands/run';
+import { registerWatchCommand } from './cli/commands/watch';
+import { registerValidateCommand } from './cli/commands/validate';
+import { registerListCommand } from './cli/commands/list';
+import { registerInitCommand } from './cli/commands/init';
+import { registerHelpCommand } from './cli/commands/help';
 import { safeResolvePath, CLIPathError } from './cli-path-utils';
 
 // Load environment variables from .env file if it exists
@@ -27,707 +28,10 @@ try {
 
 const program = new Command();
 
-// Enhanced error handling
-class CLIError extends Error {
-  constructor(message: string, public code?: string) {
-    super(message);
-    this.name = 'CLIError';
-  }
-}
-
-// Progress indicator helper
-function createProgressBar(total: number, description: string): SingleBar {
-  return new SingleBar({
-    format: chalk.blue(description) + ' |{bar}| {percentage}% | {value}/{total} | ETA: {eta}s',
-    barCompleteChar: '\u2588',
-    barIncompleteChar: '\u2591',
-    hideCursor: true
-  }, Presets.rect);
-}
-
-// Enhanced logging with colors
-function logSuccess(message: string): void {
-  console.log(chalk.green('✓'), message);
-}
-
-function logError(message: string): void {
-  console.log(chalk.red('✗'), message);
-}
-
-function logWarning(message: string): void {
-  console.log(chalk.yellow('⚠'), message);
-}
-
-function logInfo(message: string): void {
-  console.log(chalk.blue('ℹ'), message);
-}
-
 program
   .name('agentic-test')
   .description('TypeScript Agentic Testing System for Electron applications')
   .version('1.0.0');
-
-program
-  .command('run')
-  .description('Run test scenarios')
-  .option('-s, --scenario <name>', 'Run specific scenario by name')
-  .option('-d, --directory <path>', 'Scenarios directory', './scenarios')
-  .option('-c, --config <file>', 'Configuration file')
-  .option('--parallel', 'Run scenarios in parallel')
-  .option('--timeout <ms>', 'Global timeout in milliseconds', '300000')
-  .action(async (options) => {
-    try {
-      logger.info('Starting Agentic Testing System');
-
-      let config = null;
-
-      // Validate and resolve user-supplied paths (prevent path traversal - issue #93)
-      try {
-        options.directory = safeResolvePath(options.directory);
-        if (options.config) {
-          options.config = safeResolvePath(options.config);
-        }
-      } catch (err) {
-        if (err instanceof CLIPathError) {
-          throw new CLIError(err.message, 'PATH_TRAVERSAL');
-        }
-        throw err;
-      }
-
-      // Load configuration if provided
-      if (options.config) {
-        try {
-          const configManager = new ConfigManager();
-          await configManager.loadFromFile(options.config);
-          config = configManager.getConfig();
-          logSuccess(`Configuration loaded from: ${options.config}`);
-        } catch (error: any) {
-          if (error instanceof CLIError) throw error;
-          throw new CLIError(`Failed to load configuration: ${error.message}`, 'CONFIG_ERROR');
-        }
-      } else {
-        // Try loading default config files
-        const defaultConfigs = [
-          'agentic-test.config.yaml',
-          'agentic-test.config.yml',
-          'agentic-test.config.json',
-          '.agentic-testrc.yaml',
-          '.agentic-testrc.yml',
-          '.agentic-testrc.json'
-        ];
-
-        for (const configFile of defaultConfigs) {
-          try {
-            await fs.access(configFile);
-            const configManager = new ConfigManager();
-            await configManager.loadFromFile(configFile);
-            config = configManager.getConfig();
-            logInfo(`Default configuration loaded from: ${configFile}`);
-            break;
-          } catch {
-            // Continue to next file
-          }
-        }
-      }
-
-      // Validate scenario directory exists
-      try {
-        await fs.access(options.directory);
-      } catch {
-        throw new CLIError(`Scenarios directory not found: ${options.directory}`, 'DIRECTORY_NOT_FOUND');
-      }
-
-      let scenarios;
-      const progressBar = createProgressBar(1, 'Loading scenarios');
-      progressBar.start(1, 0);
-
-      try {
-        if (options.scenario) {
-          // Build scenario path and validate it stays within the resolved directory.
-          // options.scenario is a name (e.g. "my-test"), not a path, but we still
-          // guard against names containing traversal sequences like "../../etc/passwd".
-          const rawScenarioPath = path.join(options.directory, `${options.scenario}.yaml`);
-          const scenarioPath = safeResolvePath(rawScenarioPath, options.directory);
-          logInfo(`Loading scenario: ${scenarioPath}`);
-          scenarios = [await ScenarioLoader.loadFromFile(scenarioPath)];
-        } else {
-          // Load all scenarios from directory
-          logInfo(`Loading scenarios from directory: ${options.directory}`);
-          scenarios = await ScenarioLoader.loadFromDirectory(options.directory);
-        }
-        progressBar.update(1);
-        progressBar.stop();
-        
-        logSuccess(`Loaded ${scenarios.length} scenario(s)`);
-        
-        if (scenarios.length === 0) {
-          logWarning('No scenarios found to execute');
-          return;
-        }
-        
-        // Import orchestrator
-        const { TestOrchestrator } = await import('./orchestrator');
-
-        // Use shared default config, with CLI timeout override
-        const testConfig: import('./models/Config').TestConfig = config || (() => {
-          const defaults = createDefaultConfig();
-          const timeoutMs = parseInt(options.timeout, 10);
-          if (!Number.isFinite(timeoutMs) || timeoutMs < 1000 || timeoutMs > 3_600_000) {
-            throw new CLIError('--timeout must be a number between 1000 and 3600000 ms', 'INVALID_TIMEOUT');
-          }
-          defaults.execution.defaultTimeout = timeoutMs;
-          defaults.execution.resourceLimits.maxExecutionTime = timeoutMs;
-          defaults.cli.defaultTimeout = timeoutMs;
-          defaults.ui.defaultTimeout = timeoutMs;
-          defaults.tui.defaultTimeout = timeoutMs;
-          return defaults;
-        })();
-
-        const orchestrator = new TestOrchestrator(testConfig);
-
-        // Pass scenarios directly to orchestrator instead of reconstructing filenames
-        // The orchestrator will use the already-loaded scenarios
-        logInfo('Executing test scenarios...');
-        const session = await orchestrator.runWithScenarios('test-suite', scenarios);
-
-        // Extract results from session
-        const passedCount = session.summary.passed;
-        const failedCount = session.summary.failed;
-        
-        // Report results with colors
-        console.log('\n' + chalk.bold('Test Execution Results:'));
-        console.log(chalk.green(`✓ Passed: ${passedCount}`));
-        console.log(chalk.red(`✗ Failed: ${failedCount}`));
-        console.log(chalk.gray(`- Total: ${scenarios.length}`));
-        
-        if (failedCount > 0) {
-          logError('Some tests failed');
-          process.exit(1);
-        } else {
-          logSuccess('All tests passed!');
-        }
-        
-      } catch (loadError) {
-        progressBar.stop();
-        throw loadError;
-      }
-      
-    } catch (error) {
-      if (error instanceof CLIError) {
-        logError(`${error.message}`);
-        if (error.code) {
-          logInfo(`Error code: ${error.code}`);
-        }
-      } else {
-        logger.error('Operation failed', { error: error instanceof Error ? error.message : String(error), stack: error instanceof Error ? error.stack : undefined });
-        logError(`Operation failed: ${error instanceof Error ? error.message : String(error)}`);
-      }
-      process.exit(1);
-    }
-  });
-
-program
-  .command('watch')
-  .description('Watch for scenario file changes and run tests')
-  .option('-d, --directory <path>', 'Scenarios directory to watch', './scenarios')
-  .option('-c, --config <file>', 'Configuration file')
-  .action(async (options) => {
-    try {
-      // Validate and resolve user-supplied paths (prevent path traversal - issue #93)
-      try {
-        options.directory = safeResolvePath(options.directory);
-        if (options.config) {
-          options.config = safeResolvePath(options.config);
-        }
-      } catch (err) {
-        if (err instanceof CLIPathError) {
-          throw new CLIError(err.message, 'PATH_TRAVERSAL');
-        }
-        throw err;
-      }
-
-      logInfo('Starting watch mode...');
-      logInfo(`Watching directory: ${chalk.cyan(options.directory)}`);
-
-      // Validate directory exists
-      try {
-        await fs.access(options.directory);
-      } catch {
-        throw new CLIError(`Watch directory not found: ${options.directory}`, 'DIRECTORY_NOT_FOUND');
-      }
-
-      let config = null;
-
-      // Load configuration if provided
-      if (options.config) {
-        try {
-          const configManager = new ConfigManager();
-          await configManager.loadFromFile(options.config);
-          config = configManager.getConfig();
-          logSuccess(`Configuration loaded from: ${options.config}`);
-        } catch (error: any) {
-          if (error instanceof CLIError) throw error;
-          throw new CLIError(`Failed to load configuration: ${error.message}`, 'CONFIG_ERROR');
-        }
-      }
-      
-      // Setup file watcher
-      const watcher = chokidar.watch(options.directory, {
-        ignored: /(^|[\/\\])\../, // ignore dotfiles
-        persistent: true,
-        ignoreInitial: true
-      });
-      
-      // Debounce mechanism to prevent multiple rapid executions
-      let timeout: NodeJS.Timeout | null = null;
-      const debounceTime = 1000; // 1 second
-      
-      const runTests = async (changedFile?: string) => {
-        if (changedFile) {
-          logInfo(`File changed: ${chalk.yellow(changedFile)}`);
-        }
-        
-        try {
-          logInfo('Running tests...');
-          
-          // Load and execute scenarios
-          const scenarios = await ScenarioLoader.loadFromDirectory(options.directory);
-          
-          if (scenarios.length === 0) {
-            logWarning('No scenarios found to execute');
-            return;
-          }
-          
-          // Run scenarios through the real orchestrator
-          const { TestOrchestrator } = await import('./orchestrator');
-
-          const testConfig: import('./models/Config').TestConfig = config || createDefaultConfig();
-
-          const orchestrator = new TestOrchestrator(testConfig);
-          const session = await orchestrator.runWithScenarios('watch', scenarios);
-          const passedCount = session.summary.passed;
-          const failedCount = session.summary.failed;
-
-          // Report results
-          console.log('\n' + chalk.bold('Watch Mode - Test Results:'));
-          console.log(chalk.green(`✓ Passed: ${passedCount}`));
-          console.log(chalk.red(`✗ Failed: ${failedCount}`));
-          console.log(chalk.gray(`- Total: ${scenarios.length}`));
-          console.log(chalk.gray(`- Time: ${new Date().toLocaleTimeString()}`));
-          console.log('');
-
-          if (failedCount > 0) {
-            logWarning('Some tests failed - watching for changes...');
-          } else {
-            logSuccess('All tests passed - watching for changes...');
-          }
-          
-        } catch (error: any) {
-          logError(`Test execution failed: ${error.message}`);
-          logInfo('Watching for changes...');
-        }
-      };
-      
-      // File change handlers
-      watcher.on('change', (filePath) => {
-        if (timeout) clearTimeout(timeout);
-        timeout = setTimeout(() => runTests(filePath), debounceTime);
-      });
-      
-      watcher.on('add', (filePath) => {
-        logInfo(`New file added: ${chalk.green(filePath)}`);
-        if (timeout) clearTimeout(timeout);
-        timeout = setTimeout(() => runTests(filePath), debounceTime);
-      });
-      
-      watcher.on('unlink', (filePath) => {
-        logInfo(`File deleted: ${chalk.red(filePath)}`);
-        if (timeout) clearTimeout(timeout);
-        timeout = setTimeout(() => runTests(), debounceTime);
-      });
-      
-      watcher.on('error', (error) => {
-        const errorMessage = error instanceof Error ? error.message : String(error);
-        logError(`Watcher error: ${errorMessage}`);
-      });
-      
-      // Run initial test execution
-      await runTests();
-      
-      logSuccess('Watch mode started. Press Ctrl+C to stop.');
-      
-      // Handle graceful shutdown
-      process.on('SIGINT', () => {
-        logInfo('Shutting down watch mode...');
-        watcher.close().then(() => {
-          logSuccess('Watch mode stopped.');
-          process.exit(0);
-        });
-      });
-      
-    } catch (error) {
-      if (error instanceof CLIError) {
-        logError(error.message);
-        if (error.code) {
-          logInfo(`Error code: ${error.code}`);
-        }
-      } else {
-        logger.error('Operation failed', { error: error instanceof Error ? error.message : String(error), stack: error instanceof Error ? error.stack : undefined });
-        logError(`Operation failed: ${error instanceof Error ? error.message : String(error)}`);
-      }
-      process.exit(1);
-    }
-  });
-
-program
-  .command('validate')
-  .description('Validate scenario files with detailed error reporting')
-  .option('-d, --directory <path>', 'Scenarios directory', './scenarios')
-  .option('-f, --file <path>', 'Single scenario file to validate')
-  .option('--strict', 'Enable strict validation mode')
-  .action(async (options) => {
-    try {
-      // Validate and resolve user-supplied paths (prevent path traversal - issue #93)
-      try {
-        if (options.file) {
-          options.file = safeResolvePath(options.file);
-        } else {
-          options.directory = safeResolvePath(options.directory);
-        }
-      } catch (err) {
-        if (err instanceof CLIPathError) {
-          throw new CLIError(err.message, 'PATH_TRAVERSAL');
-        }
-        throw err;
-      }
-
-      logInfo('Validating scenarios...');
-
-      const parser = createYamlParser({
-        strictValidation: options.strict || false
-      });
-
-      let validationResults: Array<{file: string; valid: boolean; errors: string[]}> = [];
-
-      if (options.file) {
-        try {
-          await fs.access(options.file);
-          const result = await parser.validateYamlFile(options.file);
-          validationResults.push({
-            file: options.file,
-            valid: result.valid,
-            errors: result.errors
-          });
-
-          if (result.valid) {
-            const scenario = await ScenarioLoader.loadFromFile(options.file);
-            logSuccess(`Scenario "${scenario.name}" is valid`);
-          }
-        } catch (error: any) {
-          if (error instanceof CLIError) throw error;
-          validationResults.push({
-            file: options.file,
-            valid: false,
-            errors: [error.message]
-          });
-        }
-      } else {
-        try {
-          await fs.access(options.directory);
-        } catch {
-          throw new CLIError(`Directory not found: ${options.directory}`, 'DIRECTORY_NOT_FOUND');
-        }
-        
-        // Get all YAML files in directory
-        const files = await fs.readdir(options.directory);
-        const yamlFiles = files.filter(file => 
-          file.endsWith('.yaml') || file.endsWith('.yml')
-        );
-        
-        if (yamlFiles.length === 0) {
-          logWarning('No YAML files found to validate');
-          return;
-        }
-        
-        const progressBar = createProgressBar(yamlFiles.length, 'Validating files');
-        progressBar.start(yamlFiles.length, 0);
-        
-        for (let index = 0; index < yamlFiles.length; index++) {
-          const file = yamlFiles[index];
-          const filePath = path.join(options.directory, file);
-          try {
-            const result = await parser.validateYamlFile(filePath);
-            validationResults.push({
-              file: filePath,
-              valid: result.valid,
-              errors: result.errors
-            });
-            
-            if (result.valid) {
-              await ScenarioLoader.loadFromFile(filePath); // Additional validation
-            }
-          } catch (error: any) {
-            validationResults.push({
-              file: filePath,
-              valid: false,
-              errors: [error.message]
-            });
-          }
-          progressBar.update(index + 1);
-        }
-        progressBar.stop();
-      }
-      
-      // Report validation results
-      const validFiles = validationResults.filter(r => r.valid);
-      const invalidFiles = validationResults.filter(r => !r.valid);
-      
-      console.log('\n' + chalk.bold('Validation Results:'));
-      console.log(chalk.green(`✓ Valid files: ${validFiles.length}`));
-      console.log(chalk.red(`✗ Invalid files: ${invalidFiles.length}`));
-      
-      if (invalidFiles.length > 0) {
-        console.log('\n' + chalk.red('Validation Errors:'));
-        invalidFiles.forEach(result => {
-          console.log(chalk.red(`\n✗ ${result.file}:`));
-          result.errors.forEach(error => {
-            console.log(chalk.red(`  - ${error}`));
-          });
-        });
-        process.exit(1);
-      } else {
-        logSuccess(`All ${validationResults.length} file(s) are valid`);
-      }
-      
-    } catch (error) {
-      if (error instanceof CLIError) {
-        logError(error.message);
-        if (error.code) {
-          logInfo(`Error code: ${error.code}`);
-        }
-      } else {
-        logger.error('Operation failed', { error: error instanceof Error ? error.message : String(error), stack: error instanceof Error ? error.stack : undefined });
-        logError(`Operation failed: ${error instanceof Error ? error.message : String(error)}`);
-      }
-      process.exit(1);
-    }
-  });
-
-program
-  .command('list')
-  .description('List available scenarios with enhanced formatting')
-  .option('-d, --directory <path>', 'Scenarios directory', './scenarios')
-  .option('--json', 'Output as JSON')
-  .option('--filter <tag>', 'Filter by tag')
-  .action(async (options) => {
-    try {
-      // Validate and resolve user-supplied paths (prevent path traversal - issue #93)
-      try {
-        options.directory = safeResolvePath(options.directory);
-      } catch (err) {
-        if (err instanceof CLIPathError) {
-          throw new CLIError(err.message, 'PATH_TRAVERSAL');
-        }
-        throw err;
-      }
-
-      // Validate directory exists
-      try {
-        await fs.access(options.directory);
-      } catch {
-        throw new CLIError(`Directory not found: ${options.directory}`, 'DIRECTORY_NOT_FOUND');
-      }
-
-      const scenarios = await ScenarioLoader.loadFromDirectory(options.directory);
-      
-      if (scenarios.length === 0) {
-        logWarning('No scenarios found');
-        return;
-      }
-      
-      // Filter by tag if specified
-      let filteredScenarios = scenarios;
-      if (options.filter) {
-        filteredScenarios = scenarios.filter(scenario => 
-          scenario.metadata?.tags?.includes(options.filter) || false
-        );
-        
-        if (filteredScenarios.length === 0) {
-          logWarning(`No scenarios found with tag: ${options.filter}`);
-          return;
-        }
-      }
-      
-      if (options.json) {
-        console.log(JSON.stringify(filteredScenarios.map(scenario => ({
-          name: scenario.name,
-          description: scenario.description,
-          tags: scenario.metadata?.tags || []
-        })), null, 2));
-      } else {
-        console.log('\n' + chalk.bold('Available Scenarios:'));
-        if (options.filter) {
-          console.log(chalk.gray(`Filtered by tag: ${options.filter}`));
-        }
-        console.log();
-        
-        filteredScenarios.forEach((scenario, index) => {
-          const statusIcon = chalk.green('●'); // Always enabled for now
-          
-          console.log(`${statusIcon} ${chalk.bold(`${index + 1}. ${scenario.name}`)}`);
-          
-          if (scenario.description) {
-            console.log(`   ${chalk.gray('Description:')} ${scenario.description}`);
-          }
-          
-          if (scenario.metadata?.tags && scenario.metadata.tags.length > 0) {
-            const tagStr = scenario.metadata.tags.map(tag => chalk.cyan(`#${tag}`)).join(' ');
-            console.log(`   ${chalk.gray('Tags:')} ${tagStr}`);
-          }
-          
-          console.log();
-        });
-        
-        // Summary
-        const enabled = filteredScenarios.length;
-        const disabled = 0; // All enabled for now
-        
-        console.log(chalk.bold('Summary:'));
-        console.log(`${chalk.green('●')} Enabled: ${enabled}`);
-        console.log(`${chalk.red('○')} Disabled: ${disabled}`);
-        console.log(`Total: ${filteredScenarios.length}`);
-      }
-      
-    } catch (error) {
-      if (error instanceof CLIError) {
-        logError(error.message);
-        if (error.code) {
-          logInfo(`Error code: ${error.code}`);
-        }
-      } else {
-        logger.error('Operation failed', { error: error instanceof Error ? error.message : String(error), stack: error instanceof Error ? error.stack : undefined });
-        logError(`Operation failed: ${error instanceof Error ? error.message : String(error)}`);
-      }
-      process.exit(1);
-    }
-  });
-
-program
-  .command('init')
-  .description('Initialize a new testing project with templates and configuration')
-  .option('-d, --directory <path>', 'Target directory', './agentic-testing')
-  .option('--force', 'Overwrite existing files')
-  .option('--template <type>', 'Project template (basic, advanced, electron)', 'basic')
-  .action(async (options) => {
-    try {
-      // Validate and resolve user-supplied paths (prevent path traversal - issue #93)
-      try {
-        options.directory = safeResolvePath(options.directory);
-      } catch (err) {
-        if (err instanceof CLIPathError) {
-          throw new CLIError(err.message, 'PATH_TRAVERSAL');
-        }
-        throw err;
-      }
-
-      logInfo(`Initializing new testing project in: ${chalk.cyan(options.directory)}`);
-
-      // Check if directory exists
-      let directoryExists = false;
-      try {
-        await fs.access(options.directory);
-        directoryExists = true;
-      } catch {
-        // Directory doesn't exist, we'll create it
-      }
-      
-      if (directoryExists && !options.force) {
-        // Check if directory is empty
-        const files = await fs.readdir(options.directory);
-        if (files.length > 0) {
-          throw new CLIError(
-            `Directory ${options.directory} is not empty. Use --force to overwrite existing files.`,
-            'DIRECTORY_NOT_EMPTY'
-          );
-        }
-      }
-      
-      // Create directory if it doesn't exist
-      if (!directoryExists) {
-        await fs.mkdir(options.directory, { recursive: true });
-        logSuccess(`Created directory: ${options.directory}`);
-      }
-      
-      // Create subdirectories
-      const subdirs = ['scenarios', 'scripts', 'reports', 'screenshots', 'temp'];
-      for (const subdir of subdirs) {
-        const dirPath = path.join(options.directory, subdir);
-        await fs.mkdir(dirPath, { recursive: true });
-        logSuccess(`Created directory: ${subdir}/`);
-      }
-      
-      // Create configuration file
-      const configTemplate = getConfigTemplate(options.template);
-      const configPath = path.join(options.directory, 'agentic-test.config.yaml');
-      await fs.writeFile(configPath, configTemplate, 'utf-8');
-      logSuccess('Created configuration file: agentic-test.config.yaml');
-      
-      // Create .env template
-      const envTemplate = getEnvTemplate();
-      const envPath = path.join(options.directory, '.env.example');
-      await fs.writeFile(envPath, envTemplate, 'utf-8');
-      logSuccess('Created environment template: .env.example');
-      
-      // Create scenario templates
-      const scenarioTemplates = getScenarioTemplates(options.template);
-      for (const [filename, content] of Object.entries(scenarioTemplates)) {
-        const scenarioPath = path.join(options.directory, 'scenarios', filename);
-        await fs.writeFile(scenarioPath, content, 'utf-8');
-        logSuccess(`Created scenario template: scenarios/${filename}`);
-      }
-      
-      // Create package.json for project
-      const packageTemplate = getPackageJsonTemplate(path.basename(options.directory));
-      const packagePath = path.join(options.directory, 'package.json');
-      await fs.writeFile(packagePath, packageTemplate, 'utf-8');
-      logSuccess('Created package.json');
-      
-      // Create README.md
-      const readmeTemplate = getReadmeTemplate(options.template);
-      const readmePath = path.join(options.directory, 'README.md');
-      await fs.writeFile(readmePath, readmeTemplate, 'utf-8');
-      logSuccess('Created README.md');
-      
-      // Create .gitignore
-      const gitignoreTemplate = getGitignoreTemplate();
-      const gitignorePath = path.join(options.directory, '.gitignore');
-      await fs.writeFile(gitignorePath, gitignoreTemplate, 'utf-8');
-      logSuccess('Created .gitignore');
-      
-      console.log();
-      logSuccess('Project initialization completed!');
-      console.log();
-      console.log(chalk.bold('Next steps:'));
-      console.log(chalk.gray('1.'), `cd ${options.directory}`);
-      console.log(chalk.gray('2.'), 'cp .env.example .env');
-      console.log(chalk.gray('3.'), 'npm install');
-      console.log(chalk.gray('4.'), 'Edit scenarios in the scenarios/ directory');
-      console.log(chalk.gray('5.'), 'Run tests with: agentic-test run');
-      console.log();
-      
-    } catch (error) {
-      if (error instanceof CLIError) {
-        logError(error.message);
-        if (error.code) {
-          logInfo(`Error code: ${error.code}`);
-        }
-      } else {
-        logger.error('Operation failed', { error: error instanceof Error ? error.message : String(error), stack: error instanceof Error ? error.stack : undefined });
-        logError(`Operation failed: ${error instanceof Error ? error.message : String(error)}`);
-      }
-      process.exit(1);
-    }
-  });
 
 // Global options
 program
@@ -737,12 +41,12 @@ program
   .option('--env <file>', 'Load environment variables from file', '.env')
   .hook('preAction', (thisCommand) => {
     const opts = thisCommand.opts();
-    
+
     // Handle color option
     if (opts.noColor) {
       chalk.level = 0; // Disable colors
     }
-    
+
     // Load additional .env file if specified
     if (opts.env && opts.env !== '.env') {
       try {
@@ -758,7 +62,7 @@ program
         }
       }
     }
-    
+
     // Set logging level (simplified for compatibility)
     if (opts.debug) {
       console.log('Debug logging enabled');
@@ -767,40 +71,13 @@ program
     }
   });
 
-// Add help command with examples
-program
-  .command('help [command]')
-  .description('Display help for a specific command')
-  .action((command) => {
-    if (command) {
-      program.help({ error: false });
-    } else {
-      console.log();
-      console.log(chalk.bold('Agentic Testing System - Examples:'));
-      console.log();
-      console.log(chalk.gray('# Run all scenarios'));
-      console.log('agentic-test run');
-      console.log();
-      console.log(chalk.gray('# Run specific scenario'));
-      console.log('agentic-test run --scenario my-test');
-      console.log();
-      console.log(chalk.gray('# Run with custom configuration'));
-      console.log('agentic-test run --config my-config.yaml');
-      console.log();
-      console.log(chalk.gray('# Watch for file changes'));
-      console.log('agentic-test watch --directory ./my-scenarios');
-      console.log();
-      console.log(chalk.gray('# Initialize new project'));
-      console.log('agentic-test init --template electron --directory my-project');
-      console.log();
-      console.log(chalk.gray('# Validate scenarios'));
-      console.log('agentic-test validate --strict');
-      console.log();
-      console.log(chalk.gray('# List scenarios with filtering'));
-      console.log('agentic-test list --filter integration');
-      console.log();
-    }
-  });
+// Register all commands
+registerRunCommand(program);
+registerWatchCommand(program);
+registerValidateCommand(program);
+registerListCommand(program);
+registerInitCommand(program);
+registerHelpCommand(program);
 
 // Handle unknown commands
 program
@@ -812,401 +89,14 @@ program
     process.exit(1);
   });
 
-// Template generation functions
-function getConfigTemplate(template: string): string {
-  return `# Agentic Testing System Configuration
-# Template: ${template}
-
-execution:
-  maxParallel: ${template === 'advanced' ? 5 : 3}
-  defaultTimeout: 30000
-  continueOnFailure: true
-  maxRetries: 2
-  retryDelay: 1000
-
-cli:
-  executablePath: ${template === 'electron' ? 'npm run electron' : 'atg'}
-  workingDirectory: ${process.cwd()}
-  defaultTimeout: 30000
-  environment: {}
-  captureOutput: true
-
-ui:
-  browser: chromium
-  headless: false
-  viewport:
-    width: 1280
-    height: 720
-  baseUrl: ${template === 'electron' ? 'http://localhost:3000' : 'http://localhost:8080'}
-  defaultTimeout: 30000
-  screenshotDir: ./screenshots
-
-logging:
-  level: info
-  console: true
-  format: structured
-
-reporting:
-  outputDir: ./reports
-  formats:
-    - html
-    - json
-  includeScreenshots: true
-  includeLogs: true
-`;
-}
-
-function getEnvTemplate(): string {
-  return `# Environment Variables for Agentic Testing
-
-# Azure Configuration
-AZURE_TENANT_ID=your-tenant-id
-AZURE_CLIENT_ID=your-client-id
-AZURE_CLIENT_SECRET=your-client-secret
-
-# Application Configuration
-ELECTRON_APP_PATH=./dist/main.js
-WEBSOCKET_URL=ws://localhost:3001
-TEST_DATA_DIR=./test-data
-
-# GitHub Integration (optional)
-GITHUB_TOKEN=your-github-token
-GITHUB_OWNER=your-username
-GITHUB_REPO=your-repository
-
-# Logging
-LOG_LEVEL=info
-
-# Agentic CLI Environment Variables
-AGENTIC_LOG_LEVEL=info
-AGENTIC_MAX_PARALLEL=3
-AGENTIC_TIMEOUT=30000
-AGENTIC_HEADLESS=false
-AGENTIC_BROWSER=chromium
-AGENTIC_BASE_URL=http://localhost:3000
-AGENTIC_CLI_PATH=atg
-AGENTIC_WORKING_DIR=.
-AGENTIC_SCREENSHOT_DIR=./screenshots
-AGENTIC_REPORT_DIR=./reports
-`;
-}
-
-function getScenarioTemplates(template: string): Record<string, string> {
-  const templates: Record<string, string> = {};
-  
-  // Basic example
-  templates['example-basic.yaml'] = `id: example-basic
-name: Basic Example Test
-description: A simple test scenario to demonstrate basic functionality
-priority: MEDIUM
-interface: CLI
-prerequisites: []
-tags: [example, basic]
-steps:
-  - action: run
-    target: version
-    description: Check version
-    expected: success
-verifications:
-  - type: output
-    target: stdout
-    expected: "version"
-    operator: contains
-    description: Version output should contain version info
-estimatedDuration: 30
-enabled: true
-`;
-  
-  if (template === 'electron' || template === 'advanced') {
-    templates['example-ui.yaml'] = `id: example-ui
-name: UI Interaction Test
-description: Test user interface interactions
-priority: HIGH
-interface: UI
-prerequisites: []
-tags: [ui, interaction, ${template}]
-steps:
-  - action: click
-    target: "#start-button"
-    description: Click the start button
-    waitFor: 1000
-  - action: type
-    target: "#input-field"
-    value: "test input"
-    description: Enter test data
-  - action: click
-    target: "#submit-button"
-    description: Submit the form
-    waitFor: 2000
-verifications:
-  - type: element
-    target: "#result-message"
-    expected: "Success"
-    operator: contains
-    description: Success message should appear
-estimatedDuration: 60
-enabled: true
-`;
-  }
-  
-  if (template === 'advanced') {
-    templates['example-integration.yaml'] = `id: example-integration
-name: Integration Test Suite
-description: Complex integration test with multiple steps
-priority: CRITICAL
-interface: CLI
-prerequisites: ["database-setup", "api-server"]
-tags: [integration, api, database, advanced]
-environment:
-  TEST_MODE: integration
-  API_URL: http://localhost:3000/api
-steps:
-  - action: run
-    target: "setup"
-    description: Setup test environment
-    timeout: 10000
-  - action: run
-    target: "test-api"
-    description: Test API endpoints
-    timeout: 30000
-  - action: run
-    target: "test-database"
-    description: Test database operations
-    timeout: 15000
-verifications:
-  - type: output
-    target: stdout
-    expected: "All tests passed"
-    operator: contains
-    description: All integration tests should pass
-  - type: file
-    target: "./reports/integration-results.json"
-    expected: true
-    operator: exists
-    description: Integration report should be generated
-cleanup:
-  - action: run
-    target: "cleanup"
-    description: Clean up test environment
-estimatedDuration: 180
-enabled: true
-`;
-  }
-  
-  return templates;
-}
-
-function getPackageJsonTemplate(projectName: string): string {
-  return JSON.stringify({
-    name: projectName,
-    version: '1.0.0',
-    description: 'Agentic testing project',
-    scripts: {
-      test: 'agentic-test run',
-      'test:watch': 'agentic-test watch',
-      'test:validate': 'agentic-test validate',
-      'test:list': 'agentic-test list',
-      lint: 'echo "Add your linting command here"',
-      start: 'echo "Add your start command here"'
-    },
-    dependencies: {
-      '@azure-tenant-grapher/agentic-testing': '^1.0.0'
-    },
-    devDependencies: {},
-    keywords: ['testing', 'agentic', 'automation'],
-    author: '',
-    license: 'MIT'
-  }, null, 2);
-}
-
-function getReadmeTemplate(template: string): string {
-  const projectName = path.basename(process.cwd());
-  return `# ${projectName} - Agentic Testing Project
-
-This project was initialized with the Agentic Testing System using the **${template}** template.
-
-## Getting Started
-
-### Prerequisites
-
-- Node.js (>= 18.0.0)
-- npm or yarn
-
-### Installation
-
-\`\`\`bash
-npm install
-\`\`\`
-
-### Configuration
-
-1. Copy the environment template:
-   \`\`\`bash
-   cp .env.example .env
-   \`\`\`
-
-2. Edit the \`.env\` file with your configuration values.
-
-3. Modify \`agentic-test.config.yaml\` as needed for your project.
-
-### Running Tests
-
-\`\`\`bash
-# Run all scenarios
-npm test
-
-# Run specific scenario
-agentic-test run --scenario example-basic
-
-# Run in watch mode
-npm run test:watch
-
-# Validate scenarios
-npm run test:validate
-
-# List available scenarios
-npm run test:list
-\`\`\`
-
-## Project Structure
-
-- \`scenarios/\` - Test scenario definitions (YAML files)
-- \`reports/\` - Generated test reports
-- \`screenshots/\` - Screenshots captured during UI tests
-- \`scripts/\` - Custom scripts and utilities
-- \`temp/\` - Temporary files (auto-cleaned)
-
-## Writing Scenarios
-
-Scenarios are defined in YAML files in the \`scenarios/\` directory. See the example files for reference.
-
-### Scenario Structure
-
-\`\`\`yaml
-id: unique-scenario-id
-name: Human Readable Name
-description: What this scenario tests
-priority: CRITICAL | HIGH | MEDIUM | LOW
-interface: CLI | UI
-prerequisites: []
-tags: [tag1, tag2]
-steps:
-  - action: action-type
-    target: action-target
-    value: optional-value
-    description: What this step does
-verifications:
-  - type: verification-type
-    target: what-to-check
-    expected: expected-value
-    operator: comparison-operator
-    description: What this verifies
-estimatedDuration: 60  # seconds
-enabled: true
-\`\`\`
-
-## Contributing
-
-1. Add new scenarios to the \`scenarios/\` directory
-2. Validate your scenarios: \`npm run test:validate\`
-3. Run your tests: \`npm test\`
-4. Commit your changes
-
-## Troubleshooting
-
-- Check the \`reports/\` directory for detailed test results
-- Enable debug logging: \`agentic-test run --debug\`
-- Validate scenario syntax: \`agentic-test validate --strict\`
-
-For more information, see the [Agentic Testing System documentation](https://github.com/Azure/azure-tenant-grapher).
-`;
-}
-
-function getGitignoreTemplate(): string {
-  return `# Dependency directories
-node_modules/
-
-# Test outputs
-reports/
-screenshots/
-temp/
-
-# Logs
-*.log
-logs/
-
-# Environment variables
-.env
-.env.local
-.env.*.local
-
-# OS generated files
-.DS_Store
-.DS_Store?
-._*
-.Spotlight-V100
-.Trashes
-ehthumbs.db
-Thumbs.db
-
-# IDE files
-.vscode/
-.idea/
-*.swp
-*.swo
-*~
-
-# Runtime data
-pids
-*.pid
-*.seed
-*.pid.lock
-
-# Coverage directory used by tools like istanbul
-coverage/
-.nyc_output/
-
-# Build outputs
-dist/
-build/
-
-# Optional npm cache directory
-.npm
-
-# Optional eslint cache
-.eslintcache
-
-# Microbundle cache
-.rpt2_cache/
-.rts2_cache_cjs/
-.rts2_cache_es/
-.rts2_cache_umd/
-
-# TypeScript cache
-*.tsbuildinfo
-
-# Optional REPL history
-.node_repl_history
-
-# Output of 'npm pack'
-*.tgz
-
-# Yarn Integrity file
-.yarn-integrity
-`;
-}
-
 // Enhanced error handling for uncaught exceptions
 process.on('uncaughtException', (error) => {
-  logger.error('Uncaught exception', { error: error instanceof Error ? error.message : String(error), stack: error instanceof Error ? error.stack : undefined });
   logError(`Uncaught exception: ${error instanceof Error ? error.message : String(error)}`);
   process.exit(1);
 });
 
 process.on('unhandledRejection', (reason) => {
   const message = reason instanceof Error ? reason.message : String(reason);
-  const stack = reason instanceof Error ? reason.stack : undefined;
-  logger.error('Unhandled promise rejection', { error: message, stack });
   logError(`Unhandled promise rejection: ${message}`);
   process.exit(1);
 });
